// src/web3/contracts/types/Dropper.ts
export interface Dropper extends BaseContract {
  constructor(jsonInterface: any[], address?: string, options?: ContractOptions): Dropper;
  clone(): Dropper;
  methods: {
// ...
      signer: string,
    ): NonPayableTransactionObject<void>;

    supportsInterface(interfaceId: string | number[]): NonPayableTransactionObject<boolean>;

    surrenderPoolControl(
      poolId: number | string | BN,
      terminusAddress: string,
      newPoolController: string,
    ): NonPayableTransactionObject<void>;
// ...
  };
// ...
}